Fix NoneType object error when BK in starting inventory
ZenArcane has a mode that starts you with all the big keys, maps, compasses, and enough universal small keys to open all doors. The item pool is updated to remove them, but generation fails due to NoneType item variables.

This changes how current_amount is determined to prevent a NoneType variable access and allows for generation to continue.

I haven't noticed any adverse side-effects to this, so I decided to PR this upstream.